@@ -6,38 +6,40 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.4
-  font_awesome_flutter: ^9.2.0
+  font_awesome_flutter: ^10.1.0
+  bloc: ^8.0.3
   flutter_bloc: ^8.0.1
   flutter_secure_storage: ^5.0.2
   http: ^0.13.4
-  google_fonts: ^2.3.1
-  flutter_svg: ^1.0.3
-  image_picker: ^0.8.4+11
+  google_fonts: ^3.0.1
+  flutter_svg: ^1.1.0
+  image_picker: ^0.8.5+3
   form_field_validator: ^1.1.0
   android_intent_plus: ^3.1.1
   shimmer: ^2.0.0
-  carousel_slider: ^4.0.0
-  geolocator: ^8.2.0
-  google_maps_flutter: ^2.1.2
+  carousel_slider: ^4.1.1
+  geolocator: ^8.2.1
+  google_maps_flutter: ^2.1.6
   permission_handler: ^9.2.0
   geocoding: ^2.0.4
   intl: ^0.17.0
   polyline_do: ^0.1.0
-  url_launcher: ^6.0.20
+  url_launcher: ^6.1.2
   socket_io_client: 1.0.2
-  firebase_messaging: ^11.2.11
-  firebase_core: ^1.13.1
-  flutter_local_notifications: ^9.4.0
+  firebase_messaging: ^11.4.1
+  firebase_core: ^1.17.1
+  flutter_local_notifications: ^9.5.3+1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^2.0.0
 
 flutter:
 
