@@ -23,13 +23,13 @@ class _ListAddressesPageState extends State<ListAddressesPage> with WidgetsBindi
 
   @override
   void initState() {
-      WidgetsBinding.instance!.addObserver(this);
+      WidgetsBinding.instance.addObserver(this);
      super.initState();
   }
 
   @override
   void dispose() {
-    WidgetsBinding.instance!.removeObserver(this);
+    WidgetsBinding.instance.removeObserver(this);
     super.dispose();
   }
 
